refactor: useDataset
