+++ packages/li-sdk/src/hooks/useDataset.ts
@@ -1,12 +1,11 @@
-import { useUpdateEffect } from 'ahooks';
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback } from 'react';
 import type { DatasetFilter } from '../specs';
-import { DatasetsStoreEvent } from '../state/constants';
 import type { Dataset } from '../types';
 import { isLocalDataset, isRemoteDataset } from '../utils';
 import {
   NOOP_LOCAL_DATASET,
   NOOP_REMOTE_DATASET,
+  useDatasetBase,
   useLocalDataset,
   useRemoteDataset,
   useStateManager,
@@ -17,33 +16,9 @@ export function useDataset<T extends Dataset = Dataset>(
   options?: { filter: DatasetFilter },
 ): [T | undefined, (data: Partial<Omit<T, 'id'>>) => void] {
   const { datasetStore } = useStateManager();
-  const [dataset, setDataset] = useState(() => datasetStore.getDatasetById(datasetId));
+  const dataset = useDatasetBase<T>(datasetId);
   const isLocal = dataset && isLocalDataset(dataset);
   const isRemote = dataset && isRemoteDataset(dataset);
-
-  useUpdateEffect(() => {
-    setDataset(datasetStore.getDatasetById(datasetId));
-  }, [datasetStore.getDatasets()]);
-
-  useUpdateEffect(() => {
-    if (dataset?.id !== datasetId) {
-      setDataset(datasetStore.getDatasetById(datasetId));
-    }
-  }, [datasetId]);
-
-  useEffect(() => {
-    const onUpdateDataset = (data: Dataset) => {
-      if (data.id === datasetId) {
-        setDataset(data);
-      }
-    };
-    datasetStore.on(DatasetsStoreEvent.ADD_DATASET, onUpdateDataset);
-    datasetStore.on(DatasetsStoreEvent.UPDATE_DATASET, onUpdateDataset);
-    return () => {
-      datasetStore.off(DatasetsStoreEvent.ADD_DATASET, onUpdateDataset);
-      datasetStore.off(DatasetsStoreEvent.UPDATE_DATASET, onUpdateDataset);
-    };
-  }, [datasetId, datasetStore]);
 
   const updateDataset = useCallback(
     (data: Partial<Omit<T, 'id'>>) => {
